requirements.txt
streamlit>=1.52.0
google-genai
python-dotenv